requirements.txt
streamlit==0.51.0
numpy==1.18.1
pandas==1.0.0
scikit-learn==0.22.1
dtreeviz==0.6
psycopg2==2.8.4
python-dotenv==0.12.0
